Move setup function
